Start pulling tasks from the backend at http://todo-aerogear.rhcloud.com/todo-server.
First straw-man sketch of a) network code, and b) gson-based
serialization.  Using a simple IntentService to do networking in the
background, and a broadcast Intent to alert when a GET is complete.

NOTE: Right now we're using a static member of ToDoListFragment as the
task list storage.  This is not best practice, but does the job just
so we can stand up this toy example to explore APIs.  We're going to
probably want to use a local DB, or possibly make the data classes
Parcelable so we can pass them around in Intents.